packet: consolidate external bgp tests into src/bgp.rs
The seven external integration test files under packet/tests/bgp_*.rs had no
reason to be separate: they only used the public API and contained no
cross-crate test infrastructure. Keeping them outside src/ made navigation
harder and split related test code from the module it exercises.

Move all 127 tests into named #[cfg(test)] modules inside packet/src/bgp.rs:
nlri_tests, attribute_tests, capability_tests, framer_tests, message_tests,
misc_tests, open_tests, update_tests. The seven bare #[test] functions that
were already inline in bgp.rs are gathered into the new nlri_tests module,
which also gains access to the private BgpReader/UpdateCtx types they needed.

The 112-byte ipv6-update.raw fixture is inlined as a &[u8] literal in
parse_ipv6_update, removing the only file I/O dependency from the packet
unit tests.
