--- packet/tests/bgp_message.rs
@@ -1,295 +1,0 @@
-// Copyright (C) 2019-2024 The RustyBGP Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
-// implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use bytes::BytesMut;
-use rustybgp_packet::bgp::{
-    Attribute, Capability, Ipv4Net, Ipv6Net, Message, ParsedMessage, ParsedUpdate, PeerCodec,
-    Update,
-};
-use rustybgp_packet::{Family, Nlri, PathNlri};
-use std::collections::HashSet;
-use std::net::{Ipv4Addr, Ipv6Addr};
-use std::sync::Arc;
-
-#[test]
-fn ipv6_eor() {
-    let mut buf = [0xff; 16].to_vec();
-    let mut body: Vec<u8> = vec![
-        0x00, 0x1e, 0x02, 0x00, 0x00, 0x00, 0x07, 0x90, 0x0f, 0x00, 0x03, 0x00, 0x02, 0x01,
-    ];
-    buf.append(&mut body);
-    let mut codec = {
-        let mut c = PeerCodec::new();
-        c.set_family(Family::IPV6, Default::default());
-        c
-    };
-    assert!(codec.parse_message(&buf).is_ok());
-}
-
-#[test]
-fn parse_ipv6_update() {
-    use std::io::Read;
-    let filename = concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/tests/fixtures/ipv6-update.raw"
-    );
-    let mut file = std::fs::File::open(filename).unwrap();
-    let mut buf = Vec::new();
-    file.read_to_end(&mut buf).unwrap();
-
-    let expected: Vec<PathNlri> = vec![
-        Nlri::V6(Ipv6Net {
-            addr: Ipv6Addr::new(0x2003, 0xde, 0x2016, 0x127, 0, 0, 0, 0),
-            mask: 64,
-        }),
-        Nlri::V6(Ipv6Net {
-            addr: Ipv6Addr::new(0x2003, 0xde, 0x2016, 0x124, 0, 0, 0, 0),
-            mask: 64,
-        }),
-        Nlri::V6(Ipv6Net {
-            addr: Ipv6Addr::new(0x2003, 0xde, 0x2016, 0x128, 0, 0, 0, 0),
-            mask: 63,
-        }),
-        Nlri::V6(Ipv6Net {
-            addr: Ipv6Addr::new(0x2003, 0xde, 0x2016, 0x1ff, 0, 0, 0, 0x12),
-            mask: 127,
-        }),
-    ]
-    .into_iter()
-    .map(PathNlri::new)
-    .collect();
-
-    let mut codec = {
-        let mut c = PeerCodec::new();
-        c.set_family(Family::IPV6, Default::default());
-        c
-    };
-    let msg = codec.parse_message(&buf).unwrap();
-    match msg {
-        ParsedMessage::Update(ParsedUpdate::Routes { mp_reach, .. }) => {
-            // IPv6 NLRI arrives via MP_REACH_NLRI attribute
-            let s = mp_reach.unwrap();
-            assert_eq!(s.family, Family::IPV6);
-            assert_eq!(s.entries.len(), expected.len());
-            for (got, want) in s.entries.iter().zip(expected.iter()) {
-                assert_eq!(got, want);
-            }
-        }
-        _ => unreachable!(),
-    }
-}
-
-#[test]
-fn build_many_v4_route() {
-    let net: Vec<Nlri> = (0..2000u16)
-        .map(|i| {
-            Nlri::V4(Ipv4Net {
-                addr: Ipv4Addr::new(10, ((0xff00 & i) >> 8) as u8, (0xff & i) as u8, 1),
-                mask: 32,
-            })
-        })
-        .collect();
-
-    let mut set: HashSet<PathNlri> = net.iter().cloned().map(PathNlri::new).collect();
-
-    let mut msg = Message::Update(Update::Reach {
-        family: Family::IPV4,
-        entries: net.iter().cloned().map(PathNlri::new).collect(),
-        nexthop: None,
-        attr: Arc::new(vec![
-            Attribute::new_with_value(Attribute::ORIGIN, 0).unwrap(),
-            Attribute::new_with_bin(Attribute::AS_PATH, vec![2, 1, 1, 0, 0, 0]).unwrap(),
-            Attribute::new_with_bin(Attribute::NEXTHOP, vec![0, 0, 0, 0]).unwrap(),
-        ]),
-    });
-
-    let codec = {
-        let mut c = PeerCodec::new();
-        c.set_family(Family::IPV4, Default::default());
-        c
-    };
-    let mut txbuf = BytesMut::with_capacity(4096);
-    let mut framer = codec;
-    framer.encode_to(&msg, &mut txbuf).unwrap();
-
-    let mut recv = Vec::new();
-    loop {
-        match framer.try_parse(&mut txbuf).expect("failed to decode") {
-            Some(ParsedMessage::Update(ParsedUpdate::Routes { reach, .. })) => {
-                recv.append(&mut reach.unwrap().entries)
-            }
-            Some(_) => {}
-            None => break,
-        }
-    }
-    assert_eq!(recv.len(), net.len());
-    for n in &recv {
-        assert!(set.remove(n));
-    }
-    assert_eq!(set.len(), 0);
-
-    msg = Message::Update(Update::Unreach {
-        family: Family::IPV4,
-        entries: net.iter().cloned().map(PathNlri::new).collect(),
-    });
-    for n in &net {
-        set.insert(PathNlri::new(n.clone()));
-    }
-
-    framer.encode_to(&msg, &mut txbuf).unwrap();
-    let mut withdrawn = Vec::new();
-    loop {
-        match framer.try_parse(&mut txbuf).expect("failed to decode") {
-            Some(ParsedMessage::Update(ParsedUpdate::Routes { unreach, .. })) => {
-                withdrawn.append(&mut unreach.unwrap().entries)
-            }
-            Some(_) => {}
-            None => break,
-        }
-    }
-    assert_eq!(withdrawn.len(), net.len());
-    for n in &withdrawn {
-        assert!(set.remove(n));
-    }
-    assert_eq!(set.len(), 0);
-}
-
-#[test]
-fn many_mp_reach() {
-    let net: Vec<Nlri> = (0..2000u128)
-        .map(|i| {
-            Nlri::V6(Ipv6Net {
-                addr: Ipv6Addr::from(i),
-                mask: 128,
-            })
-        })
-        .collect();
-
-    let mut set: HashSet<PathNlri> = net.iter().cloned().map(PathNlri::new).collect();
-
-    let msg = Message::Update(Update::Reach {
-        family: Family::IPV6,
-        entries: net.iter().cloned().map(PathNlri::new).collect(),
-        nexthop: None,
-        attr: Arc::new(vec![
-            Attribute::new_with_value(Attribute::ORIGIN, 0).unwrap(),
-            Attribute::new_with_bin(Attribute::AS_PATH, vec![2, 1, 1, 0, 0, 0]).unwrap(),
-            Attribute::new_with_bin(Attribute::NEXTHOP, (0..31).collect::<Vec<u8>>()).unwrap(),
-        ]),
-    });
-
-    let codec = {
-        let mut c = PeerCodec::new();
-        c.set_family(Family::IPV6, Default::default());
-        c
-    };
-    let mut txbuf = BytesMut::with_capacity(4096);
-    let mut framer = codec;
-    framer.encode_to(&msg, &mut txbuf).unwrap();
-
-    let mut recv = Vec::new();
-    loop {
-        match framer.try_parse(&mut txbuf).expect("failed to decode") {
-            Some(ParsedMessage::Update(ParsedUpdate::Routes { mp_reach, .. })) => {
-                recv.append(&mut mp_reach.unwrap().entries)
-            }
-            Some(_) => {}
-            None => break,
-        }
-    }
-    assert_eq!(recv.len(), net.len());
-    for n in &recv {
-        assert!(set.remove(n));
-    }
-    assert_eq!(set.len(), 0);
-}
-
-#[test]
-fn many_mp_unreach() {
-    let net: Vec<Nlri> = (0..2000u128)
-        .map(|i| {
-            Nlri::V6(Ipv6Net {
-                addr: Ipv6Addr::from(i),
-                mask: 128,
-            })
-        })
-        .collect();
-
-    let mut set: HashSet<PathNlri> = net.iter().cloned().map(PathNlri::new).collect();
-
-    let msg = Message::Update(Update::Unreach {
-        family: Family::IPV6,
-        entries: net.iter().cloned().map(PathNlri::new).collect(),
-    });
-
-    let codec = {
-        let mut c = PeerCodec::new();
-        c.set_family(Family::IPV6, Default::default());
-        c
-    };
-    let mut txbuf = BytesMut::with_capacity(4096);
-    let mut framer = codec;
-    framer.encode_to(&msg, &mut txbuf).unwrap();
-
-    let mut recv = Vec::new();
-    loop {
-        match framer.try_parse(&mut txbuf).expect("failed to decode") {
-            Some(ParsedMessage::Update(ParsedUpdate::Routes { mp_unreach, .. })) => {
-                recv.append(&mut mp_unreach.unwrap().entries)
-            }
-            Some(_) => {}
-            None => break,
-        }
-    }
-    assert_eq!(recv.len(), net.len());
-    for n in &recv {
-        assert!(set.remove(n));
-    }
-    assert_eq!(set.len(), 0);
-}
-
-#[test]
-fn negotiate_extended_message_both_sides() {
-    // Both sides advertise ExtendedMessage -> extended_length = true.
-    let cap = vec![
-        Capability::MultiProtocol(Family::IPV4),
-        Capability::FourOctetAsNumber(65001),
-        Capability::ExtendedMessage,
-    ];
-    let codec = PeerCodec::negotiate(&cap, &cap);
-    assert!(
-        codec.extended_length,
-        "extended_length must be true when both sides advertise ExtendedMessage"
-    );
-}
-
-#[test]
-fn negotiate_extended_message_one_side_only() {
-    // Only local side advertises ExtendedMessage -> extended_length = false.
-    let local = vec![
-        Capability::MultiProtocol(Family::IPV4),
-        Capability::FourOctetAsNumber(65001),
-        Capability::ExtendedMessage,
-    ];
-    let remote = vec![
-        Capability::MultiProtocol(Family::IPV4),
-        Capability::FourOctetAsNumber(65002),
-    ];
-    let codec = PeerCodec::negotiate(&local, &remote);
-    assert!(
-        !codec.extended_length,
-        "extended_length must be false when only one side advertises ExtendedMessage"
-    );
-}
